backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pydantic[email]==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
# passlib removed — using stdlib PBKDF2-SHA256 for dev speed
python-multipart==0.0.9
aiofiles==23.2.1
reportlab==4.2.0
